DrawLine3D Tool
Ok so I was having issues with the raycaster not working as expected. Wasnt hitting blocks from far enough away. So I found this little script online (credits in the file) that can draw lines in the world.

With the help of something I could see I eventually found out that rotating by globalRotation resets the scaled object to 1,1,1. So changes in the rigid_body to account for this