Interaction
Sorted out some architecture for how player should use items in the hotbar or interact with objects in the world. Changes include:

Blocks now get the deltatime themselves instead of it needing to be passed from wherever player is in the flow of the interaction

Crafting table added that can be interacted with.

Instead of player doing all the raycast calculation to determine if theres something to interact with, I've decided to pass the "Player" instance to any item being used. From here the item will be able to get details about the player to decide on what to do next. The pickaxe does what player used to do in _process, detect a "Block" in front and have it take damage