Picking and Dropping
Had to make the Drop scene more generic so that it would also work with dropping items. There is still a lot of finessing to do to make it look good. Animation that I had for the "Block Drop" just wasnt working well with the pickaxe so I nixed the animation for now.

Items and Blocks will be added to a Drop instance, waite 2 seconds, and then become monitorable for pickup from the player.

It is up to the thing being added to set the position of the drop and then QueueFree() any of the children it does not need as a drop (area3D's, rigid/static bodies, etc)