Mining Fix
With the changes to the blocks to make them more uniform, theyre scaling was off when I mined them. Before I was scaling the size of the root block node, which was 1,1,1. NOW its the mesh texture I'm scaling so we need the original value of that too smoothly scale these blocks down