Better Generation!
Since I'd like the game to be competetive I want the map to mirror. So I've added a way to duplicate a map and flip it on x or y axis

Also standerdised how the map registers what blocks go in what slots. Before it was just 0,1,2. 0 was air, 1 was stone, 2 was coal. But obviously need some way to track that so just created an enum for BlockTypes.