Fix Finding Player Mouse
So the way it was before required the "Root" of the scene to be actually named Root, but I actually need it to be whatever the object actaully is. So I learned I can call the currentScene and Not the the actual root of the GetTree() which is the window on the screen. So it goes "Root/Root/player", but the 2nd root is just the currentScene so went for that