Compass to Base
WOAH this one was fun. last night I made the art, then between taking care of amanda and the dogs (sicky :frowning), was able to work out the logic. Couple notes Ill probably add in the PR:

First off in ready, originally had it look for the root scene, and then find the player and base scenes from there. Issue is that when we're in _Ready for the compass, theres no gaurente that the base has been instantiated, since its siblings with Player. so it always ended up nullptr. Had an if statement in _Process to account for this but did like that at all. Found GetNode("../") working

had fun working out the direction of the compass, which took most of my time. First I was trying angleTo, but after trying a few different vectors related to the player and the base, I drew it out and realized I was going about it wrong. I needed a kind of TrueNorth vector in the equation...like a compass! So I made a vector same X as the player and then extend to Mathf.Inf on the Z axis. But just before I got that working, I realized the Godot function AngleToPoint affectively did that. Then I just had to work out getting the distance between player and base for that vector, happy to say I didnt have to google it like i usually have to XD