More Generator Improvements
Was reading through the PR and thought of some resource based improvements in the diffusionGenerator.  Wanted to be able to set a cell on the map to any resource, not just coal. Also needed to be able to control the frequency of the resource on the map. So created a ResourceDetails inner class to help track the these details, more could be added if I think beyond these three.