Pickaxe working with Ore
Pickaxe can damage either blocks or ore, BUT when Ive called TakeDamage, it has always used the parent function. Might just be forgetting something simple, but otherwise I'll need to make an interface with TakeDamage() and then have Block and ore inherit from that. But I really think it'll just be these two types of blocks to mine so I'ma leave it as is. Not over optimize