Crocotile WorkFlow
Alrighty, thought now would be a good time to redo the textures for the blocks. Gonna be removing all the other meshs I got to improve on homemade ones. Spent all day trying to get a work flow ironed out with crocotile. Figured out a really important skill, figured out how to make objects and center their origins. Kept messing up the origins and tried to work around it, but it was clear pretty quick that I'd need to figure out how to fix the issue in crocotile, not godot.

Added Iron to the list of resources to seed. and finally I reworked the world environment a bit and got the lantern working minimally. Can be turned on and off, but only if the lantern slot is filled with any item