Reworked Pathing System!
Figured this would need to go through a re-incarnation eventually, and I'm happy with this result. Ok, lets try to explain:

So at first, the purpose of Paths was to provide a simple way to get constant values since I needed to use that to instantiate some values. Like how right now crafting table just spits out a pickaxe using Paths.Items.UI_Data.PICKAXE. Or like how in mapGenerator I have a dictionary of all the ores and their settings for maps. Ez shmeezy, nothing crazy.

BUT, I also need the paths to instantiate both UI items and physical items. When I create a pickaxe and add it to my inventory, I need the icon and a physical pickaxe. AND with this very cleanup PR, I found a situation where simply swapping a folder name in the path wont continue to work. The coal block turns into a piece of coal, which is more of a material instead of  a block. So I moved the UI's path to be Ui_Data/Crafting_Materials/coal.tscn. SO when I mined a coal block (Objects/Blocks/Coal.tscn) and just changing the Objects folder to the UI_Data folder wouldnt work. so While theres some duplication, I think I've implemented a good way to reference the Paths either hardcoded, or by the name of the object